requirements.txt
flask>=3.0.0
flask-socketio>=5.3.0
chromadb>=0.4.0
Pillow>=10.0.0
pytesseract>=0.3.10
simple-websocket>=1.0.0
groq>=0.11.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langchain-core>=0.1.20
pymupdf>=1.23.0
sentence-transformers>=2.2.2
python-telegram-bot>=20.0
requests>=2.31.0
